requirements.txt
# Core dependencies
click>=8.0.0
requests>=2.25.0
pydantic>=1.8.0
python-dotenv>=0.19.0
rich>=12.0.0
typer>=0.7.0

# AI/ML providers
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.3.0

# Android tools
adb-shell>=0.4.0

# Configuration and utilities
pyyaml>=6.0
toml>=0.10.0
jinja2>=3.0.0

# Development tools
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
mypy>=0.991
ruff>=0.0.240

# CLI enhancements
colorama>=0.4.4
tabulate>=0.9.0